Set up 'recipe' and 'user' blueprints
