+++ project/recipes/templates/recipes/base.html
@@ -1,0 +1,15 @@
+<!doctype html>
+<title>{% block title %}{% endblock %} - Recipie</title>
+<link rel="shortcut icon" href="{{ url_for('static', filename='favicon.ico') }}">
+<nav>
+  <h1>Recipie</h1>
+</nav>
+<section class="content">
+  <header>
+    {% block header %}{% endblock %}
+  </header>
+  {% for message in get_flashed_messages() %}
+    <div class="flash">{{ message }}</div>
+  {% endfor %}
+  {% block content %}{% endblock %}
+</section>
